Resolve some requests
Resolved all errors in the console except failing to load the backend,
as I assume this will succeed in loading the backend in the future. The
commented failed import has been removed, as has the commented API call
in Navbar. The switch/case in ResultsTable has been changed to use
strings instead. <Navbar></Navbar> has been replaced with <Navbar />.